abstract shared controller method permitted_dynamic_keys to new concern file
update method to be non-destructive to incoming params

include the concern class in both controllers
