+++ app/controllers/practices_controller.rb
@@ -1,5 +1,14 @@
 class PracticesController < ApplicationController # rubocop:disable Metrics/ClassLength
-  include CropperUtils, PracticesHelper, PracticeEditorUtils, EditorSessionUtils, PracticeEditorSessionsHelper, PracticeUtils, ThreeColumnDataHelper, UsersHelper
+  include CropperUtils,
+          PracticesHelper,
+          PracticeEditorUtils,
+          EditorSessionUtils,
+          PracticeEditorSessionsHelper,
+          PracticeUtils,
+          ThreeColumnDataHelper,
+          UsersHelper,
+          InnovationControllerMethods
+
   prepend_before_action :skip_timeout, only: [:session_time_remaining]
   before_action :set_practice, only: [:show, :edit, :update, :destroy, :favorite, :instructions, :overview, :impact, :resources,
                                       :documentation, :departments, :timeline, :risk_and_mitigation, :checklist, :publication_validation,
@@ -595,37 +604,6 @@ class PracticesController < ApplicationController # rubocop:disable Metrics/Clas
                                      practice_partner_practices_attributes:  [:id, :practice_partner_id, :_destroy]
 
     )
-  end
-
-  def permitted_dynamic_keys(params)
-    return {} unless params
-
-    params.transform_keys! do |key|
-      key.match?(/^\d+$/) ? "#{key}_resource" : key
-    end
-
-    params.keys.index_with do |_key|
-      [
-        :id,
-        :link_url,
-        :attachment_file_name,
-        :description,
-        :position,
-        :resource,
-        :resource_type_label,
-        :resource_type,
-        :media_type,
-        :crop_x,
-        :crop_y,
-        :crop_w,
-        :crop_h,
-        :name,
-        :image_alt_text,
-        :attachment,
-        :_destroy,
-        :value
-      ]
-    end
   end
 
   def can_view_practice
